Update lab1.py phantom_fk function

+++ lab1.py
@@ -242,7 +242,7 @@ def phantom_fk(joint_angles: np.ndarray,
   T_e_g[0:3, 0:3] = rpyr(gimbal_angles)
 
   # Full Transformation
-  phantom_T = np.array([T_0_1, T_1_2, T_2_3, T_3_e, T_e_g], dtype=float)
+  phantom_T = np.array([T_0_1, T_1_2, T_2_3, T_3_e, T_e_g])
 
   phantom_T_0_g = T_0_1 @ T_1_2 @ T_2_3 @ T_3_e @ T_e_g
